requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
markdown-it-py==3.0.0
numpy==2.1.1
scikit-learn==1.5.2
pymupdf>=1.24.0 
#pypdf==4.3.1 --> If pymupdf doesn't work
python-pptx==0.6.23
openai==1.43.0

